% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFPS.R
\name{calcFPS}
\alias{calcFPS}
\title{Calculates the framerate of a time series, in frames per second.}
\usage{
calcFPS(dat)
}
\arguments{
\item{dat}{a matrix or dataframe, or list of matrices or dataframes, that contain a variable called \code{'time'}. Data should not not yet be resampled, e.g. by \code{\link{rubitLinearInterpolate}}}
}
\value{
a numeric: the calculated framerate, in frames per second. For a single time series, the median framerate is returned; for a list of time series, the mode of individual framerates is returned.
}
\description{
Calculates the modal number of frames per second from a matrix, dataframe or list of time series.
}
\note{
This function is intended for outputs from the 'rubitrail' package.
}
\examples{
data(weevils_raw)

### Framerate of area '08'
calcFPS(weevils_raw[['08']])

### Framerate of each individual area
sapply(weevils_raw, calcFPS)

### Modal framerate of all areas combined
calcFPS(weevils_raw)
}
\seealso{
\code{\link{rubitBasic}} and \code{\link{rubitMetrics}} for information on re-encoding a new framerate into tracking data.
}

