% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lensCorrection.R
\name{lensCorrection}
\alias{lensCorrection}
\title{Corrects lens distortion in raw X,Y-coordinates.}
\usage{
lensCorrection(m, imWidth, imHeight, a = 0, b = 0, c = 0)
}
\arguments{
\item{m}{a matrix with X,Y-coordinate data labelled 'X' and 'Y'.}

\item{imWidth, imHeight}{the width and height of the original image.}

\item{a, b, c}{parameters used to correct lens distortion.}
}
\value{
The inputted matrix with corrected X,Y-coordinates.
}
\description{
Corrects lens distortion using the formula \eqn{R = r (ar^3 + br^2 + cr + d )},
where \eqn{r} is the distance to the image center and \eqn{R} is the equivalent radius for the corrected image.
}
\note{
Ready-made undistortion parameters for particular cameras can be found online (e.g. \url{http://sourceforge.net/projects/hugin/files/PTLens\%20Database}).
Parameters can be estimated manually using a checkerboard calibration image and the undistortion feature in image manipulation programs such as ImageMagick (\url{http://imagemagick.org}), or automatically using Hugin's lens calibration tool (\url{http://wiki.panotools.org/Calibrate_lens_gui}).
See \url{http://www.imagemagick.org/Usage/lens/} for more information on lens correction.
}
\examples{
data(tenebrio_basic)

tenebrio_basic <- lensCorrection(tenebrio_basic, 640, 480, b = -0.022)

}
