% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcActivity.R
\name{rubitCalcActivity}
\alias{rubitCalcActivity}
\title{Calculates the frequency and duration of mobile / stationary phases}
\usage{
rubitCalcActivity(m, window = 1, min_speed = 0.1, simple = FALSE)
}
\arguments{
\item{m}{a matrix containing processed tracking data outputted by \code{\link{rubitBasic}}}

\item{window}{the window size used to define changes in activity, in seconds.}

\item{min_speed}{the minimum speed threshold used to define changes in activity below which no movement is inferred, in mm/second.}

\item{simple}{logical; if FALSE, more complex information on average distance, velocity and position is returned for individual runs.}
}
\value{
A new matrix containing information on runs for the inputted area.
}
\description{
Defines stationary and mobile phases by determining the number of consecutive timepoints in which movement speeds are below (stationary; FALSE) or above (mobile; TRUE) a speed threshold, \code{min_speed}. The window size, \code{window}, defines the number of consecutive timepoints required to change the activity phase, in seconds.
}
\examples{
data(tenebrio)

### Apply the function over all areas in list
sapply(tenebrio, rubitCalcActivity, window = 3, simple = TRUE)
}
\seealso{
\code{\link{rubitMetrics}} to understand the different steps of processing. This function uses the run length encoding function, \code{rle()}, from base R.
}

