% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcSpeed.R
\name{rubitCalcSpeed}
\alias{rubitCalcSpeed}
\title{Calculates speed and acceleration from processed tracking data.}
\usage{
rubitCalcSpeed(m, window = 21)
}
\arguments{
\item{m}{a matrix containing processed tracking data outputted by \code{\link{rubitBasic}}.}

\item{window}{the size of the rolling median window used to smooth speed and acceleration, in frames.}
}
\value{
The inputted matrix with additional information added for speed (in mm/s) and acceleration (in mm/s^2).
}
\description{
Calculates raw and and smoothed (using a rolling median) speed and acceleration from smoothed tracking data.
}
\examples{
data(tenebrio_basic)

### Apply the function over all areas in list
sapply(tenebrio_basic, rubitCalcSpeed, window = 1)
}
\seealso{
\code{\link{rubitMetrics}} to understand the different steps of processing.
}

