% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcTurning.R
\name{rubitCalcTurning}
\alias{rubitCalcTurning}
\title{Resamples tracking data and calculates absolute and relative turning angles.}
\usage{
rubitCalcTurning(m, resample = 1)
}
\arguments{
\item{m}{a matrix containing processed tracking data outputted by \code{\link{rubitBasic}}.}

\item{resample}{the number of seconds over which to resample tracking data and calculate a new trajectory and turning angle from.}
}
\value{
A new matrix containing information on turning angles for the inputted area.
}
\description{
Calculates turning angle between successive vectors of movement by resampling data to minimise jitter, returning a matrix containing both absolute and relative angles in degrees.
}
\examples{
data(tenebrio)
### Apply the function over all areas in list
sapply(tenebrio, rubitCalcTurning, resample = 1)
}
\seealso{
\code{\link{rubitMetrics}} to understand the different steps of processing.
}

