% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertToDF.R
\name{rubitDF}
\alias{rubitDF}
\title{Flattens rubitrail lists into a single dataframe.}
\usage{
rubitDF(l, type = c("basic", "main"))
}
\arguments{
\item{l}{a list (single result file) or list of lists (multiple results files) returned by \code{\link{rubitBasic}} or \code{\link{rubitMain}}.}

\item{type}{the type of data contained in \code{l}; \code{basic} for data returned by \code{\link{rubitBasic}} and \code{main} for data returned by \code{\link{rubitMain}}.}
}
\value{
A single dataframe for 'basic' input data. A list of three dataframes for 'main' input data, with elements named 'speeds', 'turning' and 'activity'.
}
\description{
A helper function which flattens a list, or a list of lists, returned by \code{\link{rubitBasic}} and \code{\link{rubitMain}} and converts into a dataframe for ease of further analysis.
}
\examples{
data(tenebrio)

rubitDF(tenebrio, "basic")

rubitDF(tenebrio, "main")
}
\seealso{
\code{\link{rubitBasic}} and \code{\link{rubitMain}} for information on processing and analysing tracking data.
}

