% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadFile.R
\name{rubitLoadFile}
\alias{rubitLoadFile}
\title{Reads raw data from a UbiTrail result file.}
\usage{
rubitLoadFile(FILE, verbose = FALSE)
}
\arguments{
\item{FILE}{a CSV result file outputted by UbiTrail.}

\item{verbose}{logical; if TRUE, the function will print messages at every step.}
}
\value{
A list of numerical matrices with each matrix corresponding to an area.
}
\description{
A general function to read raw tracking data and meta information from a .csv result file outputted by UbiTrail.
}
\note{
The returned list contains a numerical matrix for each area.
The attributes of list contain metadata and additional information is present in each of the attributes of each matrix.
}
\examples{
### Read UbiTrail result file, e.g.:
rubitLoadFile("Result.csv")

data(weevils)

###See general metadata:
attributes(weevils)

###See how many reads in each area:
summary(weevils)

###See informations about the area named '08':
attributes(weevils[['08']])
}
\seealso{
\code{\link{rubitBasic}} and \code{\link{rubitMain}} for more further processing and analysis of tracking data.
}

