% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{rubitPlotHeatmap}
\alias{rubitPlotHeatmap}
\title{Plot a 2D density estimate of all areas.}
\usage{
rubitPlotHeatmap(l, refImg = NA, resol = 50, h = 10,
  palet = rubitTransCol(50, 0.5))
}
\arguments{
\item{l}{a list of area matrices returned by \code{\link{rubitLoadFile}} or \code{\link{rubitBasic}}.}

\item{refImg}{the path to a reference TIFF image (see details).}

\item{resol}{the resolution of the kernel density estimate.}

\item{h}{the bandwidth ratio (see details).}

\item{palet}{a vector of colours.}
}
\description{
This function represents the relative position density of each agent during the entire experiment.
}
\note{
The list \code{l} can be obtained by \code{\link{rubitBasic}}.
\code{refImg} is an image of the same size as the video frames.
If specified (\emph{i.e.} if \code{refImg != NA}), the function will use the image as background for the plot.
An easy way to obtain such an image is to ask UbiTrail to save the first frame (by ticking the corresponding box in the GUI).
For each area, the kernel estimate density function will use a bandwidth \code{bw = W/h}, 
where \code{W} is the width of the area.
}
\examples{
data(tenebrio_basic)

rubitPlotHeatmap(tenebrio_basic)
rubitPlotHeatmap(tenebrio_basic, resol=150)

}
\seealso{
\code{\link{rubitLinearInterpolate}} for interpolating data. \code{MASS::kde2d} is used by this function.
}
