\docType{package}
\name{rubitrail}
\alias{rubitrail}
\alias{Rubitrail}
\alias{rubitrail-package}
\title{Analyse Ubitrail outputfile TODO}
\description{
  \tabular{ll}{ Package: \tab rubitrail\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2012-09-24\cr
  License: \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr }
}
\author{
  Quentin Geissmann \email{qgeissmann@gmail.com}
}
\keyword{package}

