\name{rubitRead}
\alias{rubitRead}
\title{This is a function to import data from the CSV-like output generated by Ubitrail.}
\usage{
  rubitRead(FILE)
}
\arguments{
  \item{FILE}{A string indicating the path to the input
  file.}
}
\value{
  A list of matrices. Each element refers to a single area.
  Each row in the matrices represent a read (\emph{i.e.} an
  information about the tracked agent at a given time);
}
\description{
  This is a function to import data from the CSV-like
  output generated by Ubitrail.
}
\note{
  By design, Ubitrail needs to learn to discriminate the
  agent from the background. This is performed individualy
  for each area. Although this is usually fast and
  negligible, in some conditions (\emph{i.e.} if the agent
  is static or hides) it can take some time. For a given
  area, no reads will be stored if the experiment is
  interupted before the individual model is trained.
  \emph{It is therefore perfectly possible to obtain a list
  missing some areas}. In most cases, thiss can be
  interpreted as an immobile behaviour.
}
\examples{
myResult <- rubitRead(file) # this can be slow for large data
summary(myResult) # show the number of read for each area
head(myResult[[1]]) # show the first reads of the first area
Do NOT try to print large data:
rubitRead(file) ##do NOT do that
}
\seealso{
  \code{\link{rubitGlobal}} for getting general information
}
\keyword{input}

