% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/removeOutliers.R
\name{rubitRemoveOutliers}
\alias{rubitRemoveOutliers}
\title{Removes outliers from tracking data based upon log-likelihood.}
\usage{
rubitRemoveOutliers(m, p = 0.001)
}
\arguments{
\item{m}{a numerical matrix corresponding to an area.}

\item{p}{the proportion of least likely X,Y-coordinates to remove based on log-likelihood. For example, for \code{p = 0.01}, the least likely 1\% of points will be removed.}
}
\value{
A numerical matrix of the same dimensions as m.
}
\description{
Removes outliers from an area matrix based upon the log-likelihood of tracked X,Y-coordinates.
}
\examples{
data(weevils_raw)

Remove least likely 0.01\\\% of points
w_filt <- lapply(weevils_raw, rubitRemoveOutliers, p = 0.001)
}
\seealso{
\code{\link{rubitLinearInterpolate}} to interpolate X,Y-coordinates after removing outliers.
}

