% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toDF.R
\name{rubitToDF}
\alias{rubitToDF}
\title{Flattens rubitrail lists into a single dataframe.}
\usage{
rubitToDF(LIST, basic = FALSE)
}
\arguments{
\item{LIST}{a list (single result file) or list of lists (multiple results files) returned by \code{\link{rubitBasic}} or \code{\link{rubitMetrics}}.}

\item{basic}{logical defining the type of data contained in \code{LIST}. Use FALSE for metrics data returned by \code{rubitMetrics} and TRUE for basic data returned by \code{\link{rubitBasic}}.}
}
\value{
A single dataframe for 'basic' input data. A list of three dataframes for 'metrics' input data, with elements named 'speed', 'turning' and 'activity'.
}
\description{
A helper function which flattens a list, or a list of lists, returned by \code{\link{rubitBasic}} and \code{\link{rubitMetrics}} and converts into a dataframe for ease of further analysis.
}
\examples{
data(tenebrio_basic)

tenebrio_DF <- rubitToDF(tenebrio_basic, basic=TRUE)
}
\seealso{
\code{\link{rubitBasic}} and \code{\link{rubitMetrics}} for information on processing raw trajectories and extracting metrics.
}

