\name{ubitData}
\alias{ubitData}
\title{Get data from a result file.}
\usage{
  ubitData(FILE)
}
\arguments{
  \item{FILE}{the name of the file which the data are to be
  read from.}
}
\value{
  A numerical matrix.
}
\description{
  This function is used to read data from a CSV-like result
  file generated by UbiTrail.
}
\note{
  The returned matrix has a different variable in each
  column and each row is a `read'. A read is written by
  UbiTrail every time an agent is detected. Therefore, a
  read contains information about area, territory, position
  and time.
}
\examples{
\dontrun{
da <- ubitData('Results.csv')
summary(da)
}
}
\seealso{
  \code{\link{ubitMetaData}} to get only metadata from the
  first line of a CSV-like result file.

  \code{\link{ubitLoadFile}} to read a result file into a
  list of areas (more convenient).
}

