\name{ubitLoadFile}
\alias{ubitLoadFile}
\title{Read data from a result file.}
\usage{
  ubitLoadFile(FILE, verbose = FALSE)
}
\arguments{
  \item{FILE}{the name of the file which the metadata are
  to be read from.}
}
\value{
  A list of two elements. The first element contains a
  vector of strings corresponding to global informations
  such as the width and heigth of the video and the date.
  The second element contains a matrix in wich each column
  correspond to an area and four row to describe their
  respective position and dimensions.
}
\description{
  This function is used to convert all the information
  contained in a result file generated by UbiTrail into an
  R object.
}
\note{
  The metadata is in the first line of the CSV file.  It is
  pre-formated as an R expression. Therefore, if you want
  to analyse data with another toolkit than R, you will
  need to write a small function to parse the metadata.
}
\examples{
\dontrun{
l <- ubitLoadFile('Results.csv')
attributes(l)
summary(l)
}
}
\seealso{
  \code{\link{ubitMetaData}} to get only metadata from the
  first line of a CSV-like result file and
  \code{\link{ubitData}} to get only data from a CSV-like
  result file as a matrix.
}

