\name{ubitMetaData}
\alias{ubitMetaData}
\title{Get meta-data from a result file.}
\usage{
  ubitMetaData(FILE)
}
\arguments{
  \item{FILE}{the name of the file from which the metadata
  are to be read.}
}
\value{
  A list of two elements. The first element contains a
  vector of strings corresponding to global informations
  such as the width and heigth of the video and the date.
  The second element contains a matrix in wich each column
  correspond to an area and four row to describe their
  respective position and dimentions.
}
\description{
  This function is used to get general informations about a
  CSV-like result file generated by Ubitrail.
}
\note{
  The meta-data is in the first line of the CSV file.  It
  is preformated as an R expression. Therefore, if you want
  to analyse data with another toolkit than R, you need to
  write a function to parse the meta-data.
}
\examples{
\dontrun{
me <- ubitMetaData('Results.csv')
me
}
}
\seealso{
  \code{\link{ubitData}} to get only data from a CSV-like
  result file as a matrix.

  \code{\link{ubitLoad}} to read a result file into a list
  of areas (more convenient).
}

