% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{rubitLoadFile}
\alias{rubitLoadFile}
\title{Reads raw data from a UbiTrail result file.}
\usage{
rubitLoadFile(FILE, verbose = FALSE)
}
\arguments{
\item{FILE}{a CSV result file outputted by UbiTrail.}

\item{verbose}{logical; if TRUE, the function will print messages at every step.}
}
\value{
A list of numerical matrices with each matrix corresponding to an area.
}
\description{
A general function to read raw tracking data and meta information from a .csv result file outputted by UbiTrail.
}
\note{
The returned list contains a numerical matrix for each area.
The attributes of list contain metadata and additional information is present in each of the attributes of each matrix.
}
\examples{
## Locate raw data example including with package
FILE <- system.file("extdata", "tenebrio_ubitrail.csv.gz", package = "rubitrail")

tenebrio_raw <- rubitLoadFile(FILE)

###See general metadata:
attributes(tenebrio_raw)

###See how many reads in each area:
summary(tenebrio_raw)

###See informations about the area named '08':
attributes(tenebrio_raw[['08']])

}
\seealso{
\code{\link{rubitBasic}} and \code{\link{rubitMetrics}} for more further processing and analysis of tracking data.
}
