% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPosition.R
\name{rubitPlotPosition}
\alias{rubitPlotPosition}
\title{Visualise an individual trajectory and positional information in a circular area.}
\usage{
rubitPlotPosition(m, n_radials = 1, n_slices = 1, thigmo_dist = NA,
  scale = 1, n_bootstraps = 20)
}
\arguments{
\item{m}{a matrix containing processed tracking data outputted by \code{\link{rubitBasic}}}

\item{n_radials}{the number of concentric circles to divide a circular arena into}

\item{n_slices}{the number of slices to divide a circular arena into}

\item{thigmo_dist}{the distance from the boundary perimeter defined as being central (i.e. not thigmotaxis), in mm. If thigmo_dist = NA, thigmotaxis is defined as movement in the outer 50\% of the area (i.e. > \eqn{R / sqrt(2)} from the area centre, where \eqn{R} is the radius of the area).}

\item{scale}{a numeric to calibrate the true spatial scale, in pixels per mm. At the default value, measurements are returned in pixels.}

\item{n_bootstraps}{the number of random data samples used to calculate the minimum enclosing circle defining each circular area.}
}
\value{
a plot of the divided area with the full trajectory overlaid.
}
\description{
Plots a trajectory over a circular area, dividing the area into a number of grid cells of equal size to visualise exploration, and defining an outer perimeter to visualise thigmotaxis.
}
\examples{
data(tenebrio)

### Single plot, with area divided into 96 cells
### and thigmotaxis defined in outer 20mm of area:
rubitPlotPosition(tenebrio[['05']], n_radials = 8, n_slices = 12)

### Print plots of matrices in list to PDF:
#pdf("plots.pdf")
lapply(tenebrio, rubitPlotPosition, n_radials = 8, n_slices = 12)
#dev.off()
}
\seealso{
\code{\link{rubitCalcPosition}} for details on calculating positional information.
}

