\name{ubitCalcDistance}
\alias{ubitCalcDistance}
\title{Calculate distance between each reads of an area matrix.}
\usage{
  ubitCalcDistance(m)
}
\arguments{
  \item{m}{an area matrix.}
}
\description{
  This function calculates distance between the position at
  any read and the position at the next read. This is
  achived by using Pythagora theorem.
}
\examples{
print("TODO")
}

