\name{ubitLoadFile}
\alias{ubitLoadFile}
\title{Read data from a result file.}
\usage{
  ubitLoadFile(FILE, verbose = FALSE)
}
\arguments{
  \item{FILE}{the name of the file which the metadata are
  to be read from.}
}
\description{
  This function is used to get all the information
  contained in a result file generated by Ubitrail into an
  R object.
}
\note{
  The meta-data is in the first line of the CSV file.  It
  is preformated as an R expression. Therefore, if you want
  to analyse data with another toolkit than R, you need to
  write a function to parse the meta data.
}
\examples{
\dontrun{
l <- ubitLoadFile('Results.csv')
attributes(l)
summary(l)
}
}
\seealso{
  \code{\link{ubitMetaData}} to get only meta data from a
  CSV-like result file and \code{\link{ubitData}} to get
  only data from a CSV-like result file as a matrix.
}

