\name{ubitMetaData}
\alias{ubitMetaData}
\title{Get meta-data from a result file.}
\usage{
  ubitMetaData(FILE)
}
\arguments{
  \item{FILE}{the name of the file which the metadata are
  to be read from.}
}
\description{
  This function is used to get general informations about a
  CSV-like result file generated by Ubitrail.
}
\note{
  The meta-data is in the first line of the CSV file.  It
  is preformated as an R expression. Therefore, if you want
  to analyse data with another toolkit than R, you need to
  write a function to parse the meta data.
}
\examples{
print("TODO")
}
\seealso{
  \code{\link{ubitData}} to get only data from a CSV-like
  result file as a matrix.

  \code{\link{ubitLoad}} to read a result file into a list
  of areas (more convenient).
}

