\name{ubitMetaData}
\alias{ubitMetaData}
\title{Get metadata from a result file.}
\usage{
  ubitMetaData(FILE)
}
\arguments{
  \item{FILE}{the name of the file from which the metadata
  are to be read.}
}
\value{
  A list of two elements. The first element contains a
  vector of strings corresponding to global information
  such as the width and height of the video and the date.
  The second element contains a matrix in which each column
  corresponds to an area, and four rows to describe its
  respective position and dimensions.
}
\description{
  This function is used to get general information about a
  CSV-like result file generated by UbiTrail.
}
\note{
  The metadata is in the first line of the CSV file.  It is
  pre-formated as an R expression. Therefore, if you want
  to analyse data with another toolkit than R, you need to
  write a short function to parse the metadata.
}
\examples{
\dontrun{
me <- ubitMetaData('Results.csv')
me
}
}
\seealso{
  \code{\link{ubitData}} to get only data from a CSV-like
  result file as a matrix.

  \code{\link{ubitLoad}} to read a result file into a list
  of areas (more convenient).
}

